COC-178-refactor(#111): constants -> common으로 변경
